# ENH: special: faster way for calculating comb() for exact=True