--- a/scipy/special/_comb.pyx
+++ b/scipy/special/_comb.pyx
@@ -1,29 +1,25 @@
 cdef extern from "limits.h":
-    long LONG_MAX
+    unsigned long ULONG_MAX
 
-# Compute N choose k in integer arithmetic.
-def _comb(N, k):
-    cdef:
-        object numerator, denominator
-        object j, M, nterms
-    
-    if (k > N) or (N < 0) or (k < 0):
+
+def _comb_int(N, k):
+    # Fast path with machine integers
+    try:
+        r = _comb_int_long(N, k)
+        if r != 0:
+            return r
+    except (OverflowError, TypeError):
+        pass
+
+    # Fallback
+    N = int(N)
+    k = int(k)
+
+    if k > N or N < 0 or k < 0:
         return 0
 
     M = N + 1
     nterms = min(k, N - k)
-    
-    if nterms == 0:
-        return 1
-
-    if M < pow(LONG_MAX, 1./nterms):
-        # delegate to a fully typed version unless overflow
-        # The condition is rather loose: numerator < M**nterms
-        return _comb_int_unsafe(N, k)
-    elif M < LONG_MAX:
-        # both N and k fit into a C long, (but
-        # intermediaries would overflow)
-        return _comb_interm_obj(N, k)
 
     numerator = 1
     denominator = 1
@@ -34,44 +30,27 @@ def _comb(N, k):
     return numerator // denominator
 
 
-# N and k fit into a C long, use object intermediaries
-cdef object _comb_interm_obj(long N, long k):
-    cdef:
-        object numerator, denominator
-        long j, M, nterms
-    
-    if (k > N) or (N < 0) or (k < 0):
+cdef unsigned long _comb_int_long(unsigned long N, unsigned long k):
+    """
+    Compute binom(N, k) for integers.
+    Returns 0 if error/overflow encountered.
+    """
+    cdef unsigned long val, j, M, nterms
+
+    if k > N or N == ULONG_MAX:
         return 0
 
     M = N + 1
     nterms = min(k, N - k)
-    
-    numerator = 1
-    denominator = 1
+
+    val = 1
+
     for j in xrange(1, nterms + 1):
-        numerator *= M - j
-        denominator *= j
-        
-    return numerator // denominator
+        # Overflow check
+        if val > ULONG_MAX // (M - j):
+            return 0
 
+        val *= M - j
+        val //= j
 
-# not only N and k but also intermediaries fit into a C long
-cdef long _comb_int_unsafe(long N, long k):
-    cdef:
-        long numerator, denominator
-        long j, M, nterms
-    
-    if (k > N) or (N < 0) or (k < 0):
-        return 0
-
-    M = N + 1
-    nterms = min(k, N - k)  
-    
-    numerator = 1
-    denominator = 1
-    for j in xrange(1, nterms + 1):
-        numerator *= M - j
-        denominator *= j
-        
-    return numerator // denominator
-
+    return val