Cythonize the exact=True path for scipy.special.comb
Fixes gh-3449

```
  [3d938751] [ab1daa2f]
-  560.17ns   382.63ns      0.68  special.Comb.time_comb_exact(10, 100)
-    2.11μs     1.35μs      0.64  special.Comb.time_comb_exact(1000, 10)
-    2.39μs     1.42μs      0.59  special.Comb.time_comb_exact(10000, 10)
-   28.63μs    14.42μs      0.50  special.Comb.time_comb_exact(1000, 100)
-  818.32ns   400.76ns      0.49  special.Comb.time_comb_exact(100, 100)
-  913.81ns   422.43ns      0.46  special.Comb.time_comb_exact(1000, 1)
-  814.77ns   369.88ns      0.45  special.Comb.time_comb_exact(10, 10)
-  893.73ns   378.32ns      0.42  special.Comb.time_comb_exact(10000, 1)
-  896.24ns   378.52ns      0.42  special.Comb.time_comb_exact(100, 1)
-  936.60ns   378.86ns      0.40  special.Comb.time_comb_exact(10, 1)
-   38.90μs    15.29μs      0.39  special.Comb.time_comb_exact(10000, 100)
-    1.66μs   520.53ns      0.31  special.Comb.time_comb_exact(100, 10)
SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```