@ewmoore 2016-03-17
LGTM.  Anything else or is this ready to merge?